refactor: Update import path for use-taken-lecture hook
